Proyecto final/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "prophet>=1.1.6",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "uvicorn>=0.34.2",
]
